Change texts on buttons to icons. Make tags' storing in database. Minor changes.

+++ src/ViewModels/IngredientListViewModel.cs
@@ -23,8 +23,8 @@ public class IngredientListViewModel : BaseViewModel
 		}
 	}
 
-	private Ingredient selectedIngredient;
-	public Ingredient SelectedIngredient
+	private Ingredient? selectedIngredient;
+	public Ingredient? SelectedIngredient
 	{
 		get => selectedIngredient;
 		set
@@ -129,38 +129,41 @@ public class IngredientListViewModel : BaseViewModel
 	}
 	private async void EditIngredient()
 	{
-		string name = await Shell.Current.DisplayPromptAsync(
+		if (selectedIngredient != null)
+		{
+			string name = await Shell.Current.DisplayPromptAsync(
 			"",
 			$"{LocalizationManager["MsgAddingIngredient"]}",
 			$"{LocalizationManager["Ok"]}",
 			$"{LocalizationManager["Cancel"]}",
 			$"{selectedIngredient.Name}");
-		if (name != null)
-		{
-			foreach (Ingredient ingredient in Ingredients)
+			if (name != null)
 			{
-				if (ingredient != selectedIngredient && string.Equals(ingredient.Name, name, StringComparison.OrdinalIgnoreCase))
+				foreach (Ingredient ingredient in Ingredients)
 				{
-					await Shell.Current.DisplayAlert(
-						$"{LocalizationManager["Error"]}",
-						$"{LocalizationManager["ErrorIngredientExists"]}",
-						$"{LocalizationManager["Ok"]}");
-					return;
+					if (ingredient != selectedIngredient && string.Equals(ingredient.Name, name, StringComparison.OrdinalIgnoreCase))
+					{
+						await Shell.Current.DisplayAlert(
+							$"{LocalizationManager["Error"]}",
+							$"{LocalizationManager["ErrorIngredientExists"]}",
+							$"{LocalizationManager["Ok"]}");
+						return;
+					}
 				}
-			}
-			try
-			{
-				selectedIngredient.Name = name;
-				await recipeService.UpdateIngredientAsync(selectedIngredient);
-				GetIngredientsAsync();
-			}
-			catch (Exception ex)
-			{
-				Debug.WriteLine($"Unable to update ingredient: {ex.Message}");
-				await Shell.Current.DisplayAlert(
-					LocalizationManager["Error"].ToString(),
-					ex.Message,
-					LocalizationManager["Ok"].ToString());
+				try
+				{
+					selectedIngredient.Name = name;
+					await recipeService.UpdateIngredientAsync(selectedIngredient);
+					GetIngredientsAsync();
+				}
+				catch (Exception ex)
+				{
+					Debug.WriteLine($"Unable to update ingredient: {ex.Message}");
+					await Shell.Current.DisplayAlert(
+						LocalizationManager["Error"].ToString(),
+						ex.Message,
+						LocalizationManager["Ok"].ToString());
+				}
 			}
 		}
 	}
@@ -168,26 +171,40 @@ public class IngredientListViewModel : BaseViewModel
 	public ICommand DeleteIngredientCommand => new Command(DeleteIngredient, () => selectedIngredient != null);
 	private async void DeleteIngredient()
 	{
-		bool confirmed = await Shell.Current.DisplayAlert(
+		if (selectedIngredient != null)
+		{
+			bool confirmed = await Shell.Current.DisplayAlert(
 			"",
 			$"{LocalizationManager["MsgDeletingIngredient"]}",
 			$"{LocalizationManager["Ok"]}",
 			$"{LocalizationManager["Cancel"]}");
-		if (confirmed)
-		{
-			try
+			if (confirmed)
 			{
-				await recipeService.DeleteIngredientAsync(selectedIngredient);
-				GetIngredientsAsync();
-			}
-			catch (Exception ex)
-			{
-				Debug.WriteLine($"Unable to delete ingredient: {ex.Message}");
-				await Shell.Current.DisplayAlert(
-					LocalizationManager["Error"].ToString(),
-					ex.Message,
-					LocalizationManager["Ok"].ToString());
+				try
+				{
+					await recipeService.DeleteIngredientAsync(selectedIngredient);
+					GetIngredientsAsync();
+				}
+				catch (Exception ex)
+				{
+					Debug.WriteLine($"Unable to delete ingredient: {ex.Message}");
+					await Shell.Current.DisplayAlert(
+						LocalizationManager["Error"].ToString(),
+						ex.Message,
+						LocalizationManager["Ok"].ToString());
+				}
 			}
 		}
+	}
+
+	public ICommand FinishSelectionCommand => new Command(FinishSelection);
+	public async void FinishSelection()
+	{
+		Dictionary<string, object> navParam = new();
+		if (selectedIngredient != null)
+		{
+			navParam.Add(nameof(Ingredient), selectedIngredient);
+		}
+		await Shell.Current.GoToAsync("..", navParam);
 	}
 }
